fix(stars): finalize refunds and synchronize entitlements
