--- .github/workflows/pr310-final-review-fixes.yml
@@ -1,83 +1,0 @@
-name: Validate remaining PR 310 review fixes
-
-on:
-  pull_request:
-    branches:
-      - main
-    types:
-      - opened
-      - reopened
-      - synchronize
-
-permissions:
-  contents: write
-
-concurrency:
-  group: pr310-final-review-fixes-v2
-  cancel-in-progress: true
-
-jobs:
-  patch-test-commit:
-    if: github.event.pull_request.number == 310 && github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    timeout-minutes: 25
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: hotfix/stars-payment-safety
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: npm
-
-      - name: Apply remaining review fixes
-        run: |
-          set -o pipefail
-          {
-            python3 scripts/pr310-final-review-fixes.py
-            python3 scripts/pr310-final-test-updates.py
-          } 2>&1 | tee patch.log
-
-      - name: Prepare runtime data directory
-        run: |
-          sudo mkdir -p /data
-          sudo chmod 777 /data
-
-      - name: Install dependencies
-        run: npm ci --legacy-peer-deps
-
-      - name: Build TypeScript
-        run: |
-          set -o pipefail
-          npm run build 2>&1 | tee build.log
-
-      - name: Run full Jest suite
-        run: |
-          set -o pipefail
-          timeout 360s npm test -- --runInBand --forceExit 2>&1 | tee jest.log
-
-      - name: Upload validation logs
-        if: always()
-        uses: actions/upload-artifact@v4
-        with:
-          name: pr310-final-review-logs-v2
-          path: |
-            patch.log
-            build.log
-            jest.log
-          if-no-files-found: ignore
-
-      - name: Commit validated fixes
-        shell: bash
-        run: |
-          rm -f .github/workflows/pr310-final-review-fixes.yml
-          rm -f scripts/pr310-final-review-fixes.py
-          rm -f scripts/pr310-final-test-updates.py
-          rm -f patch.log build.log jest.log
-          git config user.name "chatgpt-codex-connector[bot]"
-          git config user.email "199175422+chatgpt-codex-connector[bot]@users.noreply.github.com"
-          git add -A
-          git commit -m "fix(stars): finalize refunds and synchronize entitlements"
-          git push origin HEAD:hotfix/stars-payment-safety
